// needrelay-frontend/src/api/generated/models/publicReliefRequestResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * NeedRelay API
 * Crisis and emergency resource coordination API
 * OpenAPI spec version: 0.1.0
 */
import type { PublicReliefRequestResponseStatus } from './publicReliefRequestResponseStatus';
import type { NeedResponse } from './needResponse';
import type { OrganizerContactResponse } from './organizerContactResponse';

export interface PublicReliefRequestResponse {
  id?: string;
  title: string;
  description: string;
  locationLabel: string;
  latitude: number;
  longitude: number;
  publicSlug: string;
  status?: PublicReliefRequestResponseStatus;
  needs: NeedResponse[];
  organizationName: string;
  /** @nullable */
  organizationDescription?: string | null;
  contacts: OrganizerContactResponse[];
  createdAt?: string;
}