// needrelay-frontend/src/api/generated/models/providerType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * NeedRelay API
 * Crisis and emergency resource coordination API
 * OpenAPI spec version: 0.1.0
 */

export type ProviderType = typeof ProviderType[keyof typeof ProviderType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ProviderType = {
  PERSON: 'PERSON',
  ORGANIZATION: 'ORGANIZATION',
} as const;