// needrelay-frontend/src/api/generated/models/publicReliefRequestResponseStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * NeedRelay API
 * Crisis and emergency resource coordination API
 * OpenAPI spec version: 0.1.0
 */

export type PublicReliefRequestResponseStatus = typeof PublicReliefRequestResponseStatus[keyof typeof PublicReliefRequestResponseStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PublicReliefRequestResponseStatus = {
  ACTIVE: 'ACTIVE',
  ARCHIVED: 'ARCHIVED',
} as const;